Fixed Green spike fence model.
Inside of the unity project. the model just needs to have "unity bake axis conversion" be checked. I fixed the blender import stuff. but yeah the bake axis's conversion is for both this green fence and the chainlink one. Afaik there is no other with issues such as these two. will fix if discovered that there is more.